move from codegen to lib.rs
