+++ crates/validator/src/validator.rs
@@ -1,14 +1,14 @@
 use log::error;
 use serde::Deserialize;
 use std::{fs, io, path::PathBuf, process::Command};
-use types_elm_v3::codegen;
+use types_elm_v3::EuropassEdcCredential;
 
 use crate::manifest_dir;
 
 #[derive(Debug)]
 pub struct ValidateRequest {
     pub valid_shacl: bool,
-    pub rust_object: io::Result<codegen::EuropassEdcCredential>,
+    pub rust_object: io::Result<EuropassEdcCredential>,
 }
 
 impl ValidateRequest {
@@ -40,14 +40,14 @@ fn validate_shacl(json_file: &PathBuf) -> io::Result<bool> {
     Ok(out.status.success())
 }
 
-fn validate_rust(json_file: &PathBuf) -> io::Result<codegen::EuropassEdcCredential> {
+fn validate_rust(json_file: &PathBuf) -> io::Result<EuropassEdcCredential> {
     let json = fs::read_to_string(json_file)?;
     let mut deserializer = serde_json::Deserializer::from_str(&json);
     deserializer.disable_recursion_limit();
 
     let deserializer = serde_stacker::Deserializer::new(&mut deserializer);
 
-    let credential = codegen::EuropassEdcCredential::deserialize(deserializer);
+    let credential = EuropassEdcCredential::deserialize(deserializer);
 
     if let Err(err) = &credential {
         eprintln!("Error: {err:?}");
